actualizar

+++ src/domain/services/planificacion/EstructuraService.js
@@ -3,7 +3,7 @@
 const { ErrorApp } = require('../../lib/error');
 
 module.exports = function estructuraService (repositories, helpers, res) {
-  const { EstructuraRepository } = repositories;
+  const { EstructuraRepository, OperacionRepository } = repositories;
 
   async function findAll (params) {
     try {
@@ -17,9 +17,7 @@ module.exports = function estructuraService (repositories, helpers, res) {
   async function findOne (params) {
     try {
       const estructura = await EstructuraRepository.findOne(params);
-      if (!estructura) {
-        throw new Error('El estructura no existe');
-      }
+      if (!estructura) throw new Error('El estructura no existe');
       return estructura;
     } catch (err) {
       throw new ErrorApp(err.message, 400);
@@ -33,9 +31,6 @@ module.exports = function estructuraService (repositories, helpers, res) {
         const estructuraPadre = await EstructuraRepository.findOne({ id: datos.idEstructuraPadre });
         nivel = estructuraPadre.nivel + 1;
       }
-      if (datos.gestion.id) {
-        datos.idGestion = datos.gestion.id;
-      }
       datos.nivel = nivel;
       const estructura = await EstructuraRepository.createOrUpdate(datos);
       return estructura;
@@ -46,6 +41,10 @@ module.exports = function estructuraService (repositories, helpers, res) {
 
   async function eliminar (id) {
     try {
+      const operacion = await OperacionRepository.findAll({ idEstructura: id });
+      if (operacion.rows.length > 0) throw new Error('El estructura contiene operaciones que se deben eliminar antes de eliminar esta estructura');
+      const estructuraPadre = await EstructuraRepository.findAll({ idEstructuraPadre: id });
+      if (estructuraPadre.rows.length > 0) throw new Error('El estructura contiene estructuras que son dependientes de ella que se deben eliminar antes de eliminar esta estructura');
       const estructura = await EstructuraRepository.deleteItem(id);
       return estructura;
     } catch (err) {
